refactor camera component
