+++ client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "client",
   "version": "0.1.0",
-  "proxy": "http://localhost:8080",
+  "proxy": "http://127.0.0.1:8080",
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.11.3",
