--- client/src/components/overlays/await.jsx
@@ -1,9 +1,0 @@
-
-export default function Await(props){
-
-	return (
-
-		<Div><p>Calling</p></Div>
-		
-	)
-}
