--- client/src/components/overlays/confirmation.jsx
@@ -1,8 +1,0 @@
-export default function Confirmation(props){
-
-	return (
-
-		<Div><p> Custom Confirmation Message</p></Div>
-		
-	)
-}
